+++ packages/lesswrong/lib/vulcan-lib/utils.ts
@@ -222,15 +222,6 @@ export const decodeIntlError = (error, options = {stripped: false}) => {
 
 export const isPromise = (value: any): boolean => isFunction(get(value, 'then'));
 
-export const pluralize = (s: string): string => {
-  const plural = s.slice(-1) === 'y' ?
-    `${s.slice(0, -1)}ies` :
-    s.slice(-1) === 's' ?
-      `${s}es` :
-      `${s}s`;
-  return plural;
-};
-
 export const removeProperty = (obj: any, propertyName: string): void => {
   for(const prop in obj) {
     if (prop === propertyName){
